Finished the exercise, can create a new product

+++ BestBuyCRUDBestPracticeConsoleAPP/IProductRepository.cs
@@ -7,7 +7,7 @@ namespace BestBuyCRUDBestPracticeConsoleAPP
     public interface IProductRepository
     {
         IEnumerable<Product> GetAllProducts();
-        void CreateProduct(string name, int price, int categoryId, int onSale, string stockLevel);
+        void CreateProduct(string name, decimal price, int categoryId, int onSale, string stockLevel);
 
     }
 }
